% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addRData.R
\name{addRData}
\alias{addRData}
\title{Append to an R Data File}
\usage{
addRData(robj, filename, ...)
}
\arguments{
\item{robj}{The R object to be added to the file.}

\item{filename}{The RData file to be appended.}

\item{...}{Additional arguments to save.}
}
\value{
None; Saves an RData file
}
\description{
A pseudo-append function that loads all the contents of an R data file,
adds the object, and saves the result as a new R data file.
}
\examples{
\dontrun{addRData(x, "file.RData")}
}
\author{
James C. Toll, \email{james@jtoll.com}
}
\seealso{
\code{\link[base]{readRDS}}
}

