% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dirs.R
\name{dirs}
\alias{dirs}
\title{dirs}
\usage{
dirs(path = ".")
}
\arguments{
\item{path}{a character vector containing a single path name.}
}
\value{
a vector of directories
}
\description{
Return a vector of \strong{just} the directories in a path.
}
\examples{
\dontrun{dirs()}
}
\author{
James C. Toll, \email{james@jtoll.com}
}

