% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dirsr.R
\name{dirsr}
\alias{dirsr}
\title{dirsr: a recursive version of dirs}
\usage{
dirsr(paths = ".")
}
\arguments{
\item{paths}{a character vector containing a single path name.}
}
\value{
a vector of directories
}
\description{
Recursively return a vector of \strong{just} the directories in path.
}
\examples{
\dontrun{dirsr()}
}
\author{
James C. Toll, \email{james@jtoll.com}
}

