% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\title{files}
\usage{
files(path = ".")
}
\arguments{
\item{path}{a character vector containing a single path name.}
}
\value{
a vector of files
}
\description{
Return a vector of \strong{just} the files in a path.
}
\examples{
\dontrun{files()}
}
\author{
James C. Toll, \email{james@jtoll.com}
}

