% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_vars_in_files.R
\name{find_vars_in_files}
\alias{find_vars_in_files}
\title{Find variable names in SPSS or CSV-files}
\usage{
find_vars_in_files(path = "", subf = F, file_type = "sav",
  searchstring = "", HTML_out = F)
}
\arguments{
\item{path}{the absolute path to the file folder}

\item{subf}{should subfolders also be searched}

\item{file_type}{what file type to look in (sav or csv)}

\item{searchstring}{the searchstring}

\item{HTML_out}{should a HTML-report be produced}
}
\value{
HTML-file
}
\description{
use a search string to find variables names in sav or csv files within a
specified directory
}
