% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGists.R
\name{getGists}
\alias{getGists}
\title{Get Gists}
\usage{
getGists(username)
}
\arguments{
\item{username}{A Github username.}
}
\value{
A list of the Gist names and whether they were cloned or pulled.
}
\description{
Get all of the public gists for a particular github username.
}
\examples{
\dontrun{getGists("jtoll")}
}
\author{
James C. Toll, \email{james@jtoll.com}
}
\references{
\url{https://gist.github.com/1622504}
}

