% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsdir.R
\name{lsdir}
\alias{lsdir}
\title{List the Directories/Folders on a Path}
\usage{
lsdir(path, format = "relative", recursive = FALSE, all = FALSE)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to "the working directory "."}

\item{format}{Should the directories be returned as "fullpath", "relative", or "basename"?}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{all}{logical. Should hidden files be included?}
}
\value{
A character vector containing the names of the directories in the
specified path
}
\description{
This function produces a character vector of the names of directories on
the named path.
}
\examples{
\dontrun{lsdir(".", all = TRUE)}
}
\author{
James C. Toll, \email{james@jtoll.com}
}

