% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mkdir.R
\name{mkdir}
\alias{mkdir}
\title{Make a Directory}
\usage{
mkdir(path, find = TRUE, ...)
}
\arguments{
\item{path}{The directory path to create.}

\item{find}{A boolean.}

\item{...}{Additional arguments passed to \code{\link[base]{dir.create}}.}
}
\value{
Return the pathname that was created as a string.
}
\description{
Create a new directory, possibly using a timestamp of today's date as the directory name.
}
\examples{
\dontrun{mkdir()}
\dontrun{mkdir("newDirectory")}
}
\author{
James C. Toll, \email{james@jtoll.com}
}

