% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/noclobber.R
\name{noclobber}
\alias{noclobber}
\title{No Clobber}
\usage{
noclobber(filename, behavior = c("ask", "auto", "stop"),
  form = c("timestamp", "increment"))
}
\arguments{
\item{filename}{a character vector containing a single filename.}

\item{behavior}{either "ask", "auto", or "stop".}

\item{form}{either "timestamp" or "increment".}
}
\value{
A character vector representing a unique filename.
}
\description{
A function which attempts to prevent the errant overwriting of files by
checking for their existence and if they already exist, either asking how to
proceed, timestamping, or incrementing the filename.
}
\examples{
\dontrun{noclobber("filename.csv")}
}
\author{
James C. Toll, \email{james@jtoll.com}
}

