% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sad.R
\name{sad}
\alias{sad}
\title{sad (i.e Seek And Destroy)}
\usage{
sad(pattern, envir = .GlobalEnv, all.names = FALSE, functions = FALSE,
  ...)
}
\arguments{
\item{pattern}{a character vector containing a regular expression to be matched.}

\item{envir}{the global environment is the default.}

\item{all.names}{logical. whether to include hidden (.) files in the search.}

\item{functions}{logical. whether to include functions in the search.}

\item{...}{additional arguments passed to \code{ls}.}
}
\value{
A vector of object names matching the pattern.
}
\description{
Find objects matching the pattern in the specified environment and remove.
}
\examples{
\dontrun{sad("*-tmp")}
}
\author{
James C. Toll, \email{james@jtoll.com}
}

