% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timestamp}
\alias{timestamp}
\title{Timestamp}
\usage{
timestamp(filepath)
}
\arguments{
\item{filepath}{a character vector containing a single path name.}
}
\value{
logical; indicating whether the renaming was successful.
}
\description{
Rename a file (or directory), prepending a timestamp of the modification
time to the name of the file.
}
\examples{
\dontrun{timestamp("filename.RData")}
}
\author{
James C. Toll, \email{james@jtoll.com}
}

