% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timestampr}
\alias{timestampr}
\title{Recursive Timestamping}
\usage{
timestampr(path = ".")
}
\arguments{
\item{path}{a character vector containing a single path name.}
}
\value{
logical; indicating whether the renaming was successful.
}
\description{
Add timestamps to all the non-hidden filenames in the recursive path.
}
\examples{
timestampr()
}
\author{
James C. Toll, \email{james@jtoll.com}
}

