% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/webloc.R
\name{webloc}
\alias{webloc}
\title{Webloc}
\usage{
webloc(path = ".", recursive = FALSE)
}
\arguments{
\item{path}{a character vector containing a single path name.}

\item{recursive}{logical.}
}
\value{
A dataframe of URLs.
}
\description{
Read the .webloc files created by web browsers as URL bookmarks into a
dataframe. NOTE: This does not work with XML formatted .webloc files.
}
\examples{
\dontrun{webloc()}
}
\author{
James C. Toll, \email{james@jtoll.com}
}

